+++ scratch.js
@@ -1,9 +1,9 @@
-const informal = require('./src/index')
+const soft = require('./src/index')
 const spacetime = require('/Users/spencer/mountain/spacetime')
 // let zones = spacetime().timezones
 
-let iana = informal.find('5gmt')
-const show = informal.display(iana)
+let iana = soft.find('ist')
+const show = soft.display(iana)
 console.log(show)
 
 // const tzCodes = informal.display('Canada/Eastern')