+++ test/allwords.test.js
@@ -1,6 +1,6 @@
 const test = require('tape')
 const spacetime = require('spacetime')
-const informal = require('../src')
+const soft = require('../src')
 let zones = spacetime().timezones
 
 const skip = {
@@ -13,7 +13,7 @@ const skip = {
 let list = Object.keys(require('../data'))
 test('all words produce valid iana', (t) => {
   list.forEach((str) => {
-    let found = informal.find(str)
+    let found = soft.find(str)
     t.ok((found && zones[found.toLowerCase()]) || skip[str], str)
   })
   t.end()