+++ test/display.test.js
@@ -1,5 +1,5 @@
 const test = require('tape')
-const informal = require('../src')
+const soft = require('../src')
 
 test('display-test', (t) => {
   let arr = [
@@ -22,7 +22,7 @@ test('display-test', (t) => {
     ['darwin', 'ACST', 'ACDT']
   ]
   arr.forEach((a) => {
-    let display = informal.display(a[0])
+    let display = soft.display(a[0])
     t.equal(display.standard.abbrev, a[1], a[0] + ' standard')
     if (display.daylight) {
       t.equal(display.daylight.abbrev, a[2], a[0] + ' daylight')